api_comms/requirements.txt
adblockparser==0.7
aiodns==2.0.0
aiohttp==3.7.3
aiomultiprocess==0.8.0
aiosqlite==0.16.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.1.0
automium==0.2.6
automium-web==0.1.1
autopep8==1.5.5
backoff==1.10.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
censys==1.1.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cheroot==8.5.1
CherryPy==18.6.0
cherrypy-cors==1.6
click==7.1.2
click-plugins==1.1.1
colorama==0.4.3
company==0.0.4
coverage==5.4
crypto==1.4.1
cryptography==3.3.1
decorator==4.4.2
distlib==0.3.1
Django==3.1.7
djangorestframework==3.12.4
dnspython==2.0.0
edgar==5.4.1
essential-generators==1.0
ExifRead==2.3.2
filelock==3.0.12
flake8==3.9.0
Flask==1.1.2
future==0.18.2
fuzzywuzzy==0.18.0
html2text==2020.1.16
httpagentparser==1.9.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
ipaddr==2.2.0
ipwhois==1.0.0
isort==5.6.4
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.1.0
jaraco.functools==3.1.0
jaraco.text==3.4.0
Jinja2==2.11.3
jsonfield==3.1.0
keyboard==0.13.5
lazy-object-proxy==1.4.3
legacy==0.1.6
lxml==4.6.1
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
multidict==5.1.0
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==2.0.3
Naked==0.1.31
netaddr==0.8.0
networkx==2.5
numpy==1.20.1
packaging==20.4
phonenumbers==8.12.15
pickle-warehouse==0.1.1
Pillow==8.0.1
plotly==4.14.1
pluggy==0.13.1
portend==2.7.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
publicsuffixlist==0.7.6
py==1.9.0
pycares==3.1.1
pycodestyle==2.7.0
pycparser==2.20
pyee==7.0.4
pyflakes==2.3.1
pygexf==0.2.2
pylint==2.6.0
pynput==1.7.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.2
PySocks==1.7.1
pytest==6.0.1
pytest-cov==2.11.1
python-docx==0.8.10
python-Levenshtein==0.12.2
python-pptx==0.6.18
python-whois==0.7.3
pytz==2020.5
pywin32==300
PyYAML==5.3.1
randomtimestamp==1.0.0
regex==2020.10.15
register==0.1
requests==2.24.0
requests-futures==1.0.0
retrying==1.3.3
scapy==2.4.4
secure==0.2.1
selenium==3.141.0
shellescape==3.8.1
shodan==1.24.0
six==1.15.0
soupsieve==2.1
spotipy==2.18.0
sqlparse==0.4.1
status==0.2.5
stem==1.8.0
tempora==4.0.1
texttable==1.6.3
toml==0.10.1
torrequest==0.1.0
tqdm==4.55.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
uuid==1.30
views==0.3
virtualenv==20.4.3
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
XlsxWriter==1.3.7
yarl==1.6.3
zc.lockfile==2.0
zipp==3.1.0
